+++ resticprofile-backup/rootfs/usr/local/bin/init.sh
@@ -1,6 +1,29 @@
 #!/usr/bin/with-contenv bashio
+# shellcheck shell=bash
 
 set -euo pipefail
+
+tmp_avail="$(mktemp)"
+# Available devices
+blkid | awk '{print substr($1, 0, length($1) - 1)}' | awk -F'/' '{print $NF}' > "$tmp_avail"
+echo "NAME" >> "$tmp_avail"
+
+## List available Disk with Labels and Id
+bashio::log.blue "---------------------------------------------------"
+bashio::log.green "Available local disks for mounting:"
+output=$(lsblk -o name,label,size,fstype,ro | awk '$4 != "" { print $0 }' | grep -f "$tmp_avail")
+if [ "$(echo "$output" | grep -c '^')" -le 1 ]; then
+  bashio::log.yellow "No local disks available"
+else
+  echo "$output"
+fi
+bashio::log.blue "---------------------------------------------------"
+rm -f "$tmp_avail"
+
+# Show support fs https://github.com/dianlight/hassio-addons/blob/2e903184254617ac2484fe7c03a6e33e6987151c/sambanas/rootfs/etc/s6-overlay/s6-rc.d/init-automount/run#L106
+fstypessupport=$(grep -v nodev < /proc/filesystems | awk '{$1=" "$1}1' | tr -d '\n\t')
+bashio::log.green "Supported fs: ${fstypessupport}"
+bashio::log.blue "---------------------------------------------------"
 
 # if password file does not exist, generate a new one
 if [ ! -f /config/password.txt ]; then